Moved checkCRD to config/kubernetes.js
